requirement.txt
backoff==2.2.1
bleak==0.22.3
blinker==1.9.0
click==8.1.8
dbus-fast==2.24.4
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
RPi.GPIO==0.7.1
typing_extensions==4.12.2
Werkzeug==3.1.3
